+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Include filtering class for situation report and appeal documents.
 - Include report date, updated at timestamp in field report response.
 - Include search fields for all relevant admin forms.
+- Include alert sorting and ordering.
 
 ### Changed